+++ stormbird/src/actuator_line/corrections/empirical_circulation.rs
@@ -1,14 +1,15 @@
 
 use serde::{Deserialize, Serialize};
 
-use std::f64::consts::PI;
+use stormath::type_aliases::Float;
+use stormath::consts::PI;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EmpiricalCirculationCorrection {
     #[serde(default = "EmpiricalCirculationCorrection::default_exp_factor")]
-    pub exp_factor: f64,
+    pub exp_factor: Float,
     #[serde(default = "EmpiricalCirculationCorrection::default_one")]
-    pub overall_correction: f64,
+    pub overall_correction: Float,
 }
 
 impl Default for EmpiricalCirculationCorrection {
@@ -21,10 +22,10 @@ impl Default for EmpiricalCirculationCorrection {
 }
 
 impl EmpiricalCirculationCorrection {
-    fn default_one() -> f64 {1.0}
-    fn default_exp_factor() -> f64 {10.0}
+    fn default_one() -> Float {1.0}
+    fn default_exp_factor() -> Float {10.0}
 
-    pub fn correction_factor(&self, non_dim_span_position: f64) -> f64 {
+    pub fn correction_factor(&self, non_dim_span_position: Float) -> Float {
         let s_in = 0.5 - non_dim_span_position.abs();
 
         let exp_value = (-self.exp_factor * s_in).exp();
