+++ stormbird/src/lifting_line/singularity_elements/mod.rs
@@ -10,10 +10,12 @@ pub mod vortex_line;
 pub mod symmetry_condition;
 pub mod panel;
 
+use stormath::type_aliases::Float;
+
 #[derive(Clone, Debug)]
 /// Settings and functions to calculate induced velocities from vortex lines and vortex panels.
 pub struct PotentialTheorySettings {
-    pub far_field_ratio: f64,
+    pub far_field_ratio: Float,
     pub symmetry_condition: SymmetryCondition,
 }
 
@@ -27,7 +29,7 @@ impl Default for PotentialTheorySettings {
 }
 
 impl PotentialTheorySettings {
-    pub fn default_far_field_ratio() -> f64 { 5.0 }
+    pub fn default_far_field_ratio() -> Float { 5.0 }
 }
 
 pub mod prelude {
